drop ablations on summay again after icse27 submission

+++ src/main/scala/esmeta/analyzer/tychecker/AbsValue.scala
@@ -53,9 +53,6 @@ trait AbsValueDecl { self: TyChecker =>
       val ty = this.ty
       this.copy(guard = (ty, this.guard) && (ty, guard))
 
-    /** remove type guard */
-    def withoutTypeGuard: AbsValue = this.copy(guard = TypeGuard.Empty)
-
     /** kill bases */
     def kill(bases: Set[Base], update: Boolean)(using AbsState): AbsValue =
       val ty = this.symty.bases.exists(bases.contains) match
@@ -71,11 +68,10 @@ trait AbsValueDecl { self: TyChecker =>
       entrySt: AbsState,
     ): AbsValue =
       given AbsState = givenSt
-      val ret = if (isTypeGuardCandidate(func)) {
+      if (isTypeGuardCandidate(func)) {
         val xs = givenSt.getImprecBases(entrySt)
         this.onlySym.kill(xs, update = false)
       } else AbsValue(this.ty)
-      if (config.resultTypeInsensitive) ret.withoutTypeGuard else ret
 
     /** get symbols */
     def bases: Set[Base] =
